alacrity_backend/requirements.txt
Django
mysqlclient  
djangorestframework
pymysql
django-cors-headers
nanoid
pandas
scipy
statsmodels
duckdb
cryptography
pyarrow
seaborn
scikit-learn
tenseal

channels
channels-redis
daphne





# dotenv
python-dotenv


matplotlib

# for payment
paypalrestsdk




# storage for files
django-storages
boto3
django-storages[boto3]
minio
redis

# for minio
django-minio-storage

# authentication 
pillow
rest_framework_simplejwt

# for background tasks

# for authentication
# djangorestframework-jwt
djangorestframework-simplejwt
Pillow



django-environ









